+++ src/App.tsx
@@ -12,6 +12,7 @@ import Data from "./pages/services/Data";
 import Content from "./pages/services/Content";
 import Consulting from "./pages/services/Consulting";
 import MusicService from "./pages/services/Music";
+import NoCode from "./pages/services/NoCode";
 import Resources from "./pages/Resources";
 import Blog from "./pages/Blog";
 import BlogPost from "./pages/BlogPost";
@@ -41,6 +42,7 @@ const App = () => {
         <Route path="/services/content" element={<Content />} />
         <Route path="/services/consulting" element={<Consulting />} />
         <Route path="/services/music" element={<MusicService />} />
+        <Route path="/services/nocode" element={<NoCode />} />
         <Route path="/resources" element={<Resources />} />
         <Route path="/blog" element={<Blog />} />
         {/* Specific blog post routes should come before the dynamic route */}